Somewhat finish working on Silicon
Silicon is WINTER's exec engine.

BUG to fix:
Bug in AOs, where commands such as `pixelate "search anything"`
basically in this the string literal " is not being removed from the string.
Therefore the search query prepared by AOs is something like:
https://www.google.com/search?q="search+anything"
which is not good. Fix it in AOs 2.8 which is going to be built in C++.